+++ modules/contribute/contribute.py
@@ -7,6 +7,9 @@ from modules import util
 def generate_contribute():
     """Generate contribute page markdown"""
 
+    # Create content pages directory if does not already exist
+    util.buildhelpers.create_content_pages_dir()
+
     # Move templates to templates directory
     util.buildhelpers.move_templates(contribute_config.module_name, contribute_config.contribute_templates_path)
 
@@ -41,6 +44,10 @@ def generate_contribute():
 
     subs = contribute_config.contribute_index_md + json.dumps(data)
 
-    #Open markdown file for the contribute page
+    # Create directory if it does not exist
+    if not os.path.isdir(contribute_config.contribute_markdown_path):
+        os.mkdir(contribute_config.contribute_markdown_path)
+
+    # Open markdown file for the contribute page
     with open(os.path.join(contribute_config.contribute_markdown_path, "contribute.md"), "w", encoding='utf8') as md_file:
         md_file.write(subs)