+++ modules/resources/resources.py
@@ -1,7 +1,7 @@
 import json
 import os
 from . import resources_config
-from . import versions
+import modules
 from modules import site_config
 from modules import util
 from datetime import datetime
@@ -9,9 +9,12 @@ from datetime import datetime
 def generate_resources():
     """Responsible for generating the resources pages"""
 
+    # Create content pages directory if does not already exist
+    util.buildhelpers.create_content_pages_dir()
+
     # Verify if resources directory exists
-    if not os.path.isdir(resources_config.resources_markdown_path):
-        os.mkdir(resources_config.resources_markdown_path)
+    if not os.path.isdir(site_config.resources_markdown_path):
+        os.mkdir(site_config.resources_markdown_path)
 
     # Verify if resources directory exists
     if not os.path.isdir(resources_config.updates_markdown_path):
@@ -19,13 +22,11 @@ def generate_resources():
 
     # Move templates to templates directory
     util.buildhelpers.move_templates(resources_config.module_name, resources_config.resources_templates_path)
-
+    util.buildhelpers.move_docs(resources_config.docs_path)
     generate_general_information()
     generate_training_pages()
     generate_attackcon_page()
-    generate_faq_page()
-    generate_versions()
-    generate_changelog_page()
+    check_versions_module()
     generate_static_pages()
 
 def generate_general_information():
@@ -45,7 +46,7 @@ def generate_general_information():
         "presentations": presentations
     })
     # write markdown to file
-    with open(os.path.join(resources_config.resources_markdown_path, "general_information.md"), "w", encoding='utf8') as md_file:
+    with open(os.path.join(site_config.resources_markdown_path, "general_information.md"), "w", encoding='utf8') as md_file:
         md_file.write(resources_content)
 
 def generate_training_pages():
@@ -60,14 +61,14 @@ def generate_training_pages():
     training_md = resources_config.training_md + json.dumps(data)
 
     # write markdown to file
-    with open(os.path.join(resources_config.resources_markdown_path, "training.md"), "w", encoding='utf8') as md_file:
+    with open(os.path.join(site_config.resources_markdown_path, "training.md"), "w", encoding='utf8') as md_file:
         md_file.write(training_md)
 
     # CTI training
     training_cti_md = resources_config.training_cti_md + json.dumps(data)
 
     # write markdown to file
-    with open(os.path.join(resources_config.resources_markdown_path, "training_cti.md"), "w", encoding='utf8') as md_file:
+    with open(os.path.join(site_config.resources_markdown_path, "training_cti.md"), "w", encoding='utf8') as md_file:
         md_file.write(training_cti_md)
 
 def generate_attackcon_page():
@@ -82,44 +83,15 @@ def generate_attackcon_page():
 
     attackcon_content = resources_config.attackcon_md + json.dumps(attackcon)
     # write markdown to file
-    with open(os.path.join(resources_config.resources_markdown_path, "attackcon.md"), "w", encoding='utf8') as md_file:
+    with open(os.path.join(site_config.resources_markdown_path, "attackcon.md"), "w", encoding='utf8') as md_file:
         md_file.write(attackcon_content)
 
-def generate_faq_page():
-    """Responsible for compiling faq json into faq markdown file
-       for rendering on the HMTL
-    """
-    # load faq data from json
-    with open(os.path.join(site_config.data_directory, "faq.json"), "r", encoding='utf8') as f:
-        faqdata = json.load(f)
-    # add unique IDs
-    for i,section in enumerate(faqdata["sections"]):
-        for j,item in enumerate(section["questions"]):
-            item["id"] = f"faq-{i}-{j}"
-    
-    # get markdown
-    faq_content = resources_config.faq_md + json.dumps(faqdata)
-    # write markdown to file
-    with open(os.path.join(resources_config.resources_markdown_path, "faq.md"), "w", encoding='utf8') as md_file:
-        md_file.write(faq_content)
+def check_versions_module():
+    """ Verify if versions module is in the running pool """
 
-def generate_versions():
-    versions.deploy()
+    if not [key['name'] for key in modules.run_ptr if key['name'] == 'versions']:
+        util.buildhelpers.remove_element_from_sub_menu(resources_config.module_name, "Versions of ATT&CK")
 
-def generate_changelog_page():
-    """Responsible for compiling original changelog markdown into changelog markdown file
-       for rendering on the HTML
-    """
-    
-    # Read local changelog
-    with open("CHANGELOG.md", "r", encoding='utf8') as f:
-        changelog = f.read()
-    
-    # Append changelog to mardown file
-    changelog_md = resources_config.changelog_md + changelog
-
-    with open(os.path.join(resources_config.resources_markdown_path, "changelog.md"), "w", encoding='utf8') as md_file:
-        md_file.write(changelog_md)
 
 def generate_static_pages():
     """ Reads markdown files from the static pages directory and copies them into 
@@ -137,5 +109,5 @@ def generate_static_pages():
                 with open(os.path.join(resources_config.updates_markdown_path, static_page), "w", encoding='utf8') as md_file:
                     md_file.write(content)
             else:
-                with open(os.path.join(resources_config.resources_markdown_path, static_page), "w", encoding='utf8') as md_file:
+                with open(os.path.join(site_config.resources_markdown_path, static_page), "w", encoding='utf8') as md_file:
                     md_file.write(content)