+++ modules/search/search.py
@@ -2,7 +2,11 @@ import os
 import bleach, re
 import json
 import html
-from modules import resources
+import importlib
+import modules
+# Check if resources module exist
+if importlib.util.find_spec('modules.versions'):
+    from modules import versions
 from modules import site_config
 
 def generate_index():
@@ -98,5 +102,6 @@ def clean(filepath):
 def preserve_current_version():
     """ Preserve current version """
     
-    if resources.versions:
-        resources.versions.deploy_current_version()
+    # Check for intermodule dependency
+    if [key['name'] for key in modules.run_ptr if key['name'] == 'versions']:
+        versions.versions.deploy_current_version()