+++ CHANGELOG.md
@@ -26,7 +26,7 @@
 <!--          addition of notable features. When PATCH is 0 it can      -->
 <!--          be omitted                                                -->
 
-# Modularization
+# Changes staged on develop
 ## ATT&CK Website version 3.0
 ### Improvements
 - Added a link on the home page that takes the user to a random page within a specified category. See issue [#98](https://github.com/mitre-attack/attack-website/issues/98).
@@ -39,7 +39,9 @@
 - Matrix layouts on the home page and matrices page now persist across pages and sessions. See issue [#165](https://github.com/mitre-attack/attack-website/issues/165).
 - Added [Network matrix](https://attack.mitre.org/matrices/network). See issue [#230](https://github.com/mitre-attack/attack-website/issues/230).
 - Removed PRE-ATT&CK domain to support migration into the new (TODO: list and link to tactics) tactics in Enterprise-ATT&CK. See issue [#222](https://github.com/mitre-attack/attack-website/issues/222).
-- Added PRE platform. See issue [#251](https://github.com/mitre-attack/attack-website/issues/251).
+- Added [PRE matrix](https://attack.mitre.org/matrices/PRE). See issue [#251](https://github.com/mitre-attack/attack-website/issues/251).
+- Website built by users are visually distinct from attack.mitre.org unless brand flag is added as an argument. See issue [#240](https://github.com/mitre-attack/attack-website/issues/240).
+- Website is built without specific related ATT&CK content such as resources, contribute, and blog unless specified. See issue [#241](https://github.com/mitre-attack/attack-website/issues/241).
 
 ### Fixes
 - Fixed bug where bootstrap dropdown menu buttons require two clicks to open the first time they are opened. See issue [#152](https://github.com/mitre-attack/attack-website/issues/152).